created a Play module, and refactored test structure
the new Play module contains everything associated with putting a word on the board - validation, scoring, etc. this stuff was spread out across several files, and it makes more sense now.

refactored the tests so that each module has it's own test file. this makes it easier to see what is and isn't tested. it might be extreme, but, i don't think so yet.