a bunch of refactorings.
created a Move package, and moved Play into it (renaming it to Move)
refactored out Validation and Scoring from Move, for more manageable code
renamed PutWord to WordPut, though I still dont really like that much either
made Game depend on Move, rather than the opposite. Moves know nothing about Games and thats good.
added a list of Turns to Game, that way it can support undo. A Turn is the player, the move, and the game at the start of the players turn.
added some haddock, especially a description for every module