huge board refactoring to use arrays, and got rid of Pos
trying to have a typeclass to allow for different board representations just didn't work - it was awful code. ive replaced it all just by using an array for the board instead. this cleaned up a tremendous amount of code clutter. there are still some opportunities for optimizations, too.

 in the process, I removed Matrix and ListBoard. Board is just an Array, and its hardcoded directly into the Board module.

 got rid of all the Pos/Position crap, because it was also just clutter. just using (Int, Int) for points is way simpler. this was over-engineered because I thought i wanted to use it to teach people how to make their own typeclasses, but, just a bad example. oh well.

i also made it possible to turn validation off, so that you can place anything on the board without worrying about if it is a proper word or not. this makes it possible to test a whole bunch of other functionality, particularly scoring, more easily.

the tests still run a little slowly, so there is a perf problem somewhere that still needs to be found.